Fix icon button height calculation to ensure proper integer value

+++ TracX/ui/common/icon_button.py
@@ -30,7 +30,7 @@ class IconButton(QPushButton):
 
     def setIconFile(self, iconFile, iconSize):
         if iconFile is None:
-            self.setFixedHeight(iconSize * 1.5)
+            self.setFixedHeight(int(iconSize * 1.5))
             return
 
         path = os.path.join(APP_ASSETS, "icons", iconFile)
